requirements.txt
# Web backend
Flask==3.0.3
waitress==3.0.0

# Model inference
tensorflow-cpu==2.16.1
numpy>=1.26.0,<2.0.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0

# Local evaluation and notebooks
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scipy>=1.10.0
